==> 02_r_scripts/01_fix_flow_direction.R <==
# ------------------------------ #
### --- Fix flow direction --- ### 
# ------------------------------ #

# date written: 21.10.20
# date changed: 28.10.20 

# Used to be part of 005_create_map.R. Exacted because it is also useful for other parts of the project. 

# overview # 

## OPTIONS 



OPTIONS = list(workshop= F, 
               remove= F,
               save = T)

# setup -------------------------------------------------------------------
pacman::p_load(data.table,
               dplyr,
               here,
               lwgeom,
               magrittr,
               sf,
               stringr,
               tmap)

# create DIR if not present 
if (!str_detect("DIR", ls())){
source(textConnection(readLines(file.path("02_r_scripts/02_check_flow.R"))[28:30]))
}

# load reverse function 
source(file.path(DIR$rs, "f_01_reverse.R"))
source(file.path(DIR$rs,  "f_03_split_rivers.R"))
source(file.path(DIR$rs,  "f_04_clean_node.R"))
source(file.path(DIR$rs,  "f_05_split_rivers_external.R"))

# load data ---------------------------------------------------------------
dt_rivers = readRDS(file.path(DIR$da, "rivers_w_added.RDS"))
st_sites  = readRDS(file.path(DIR$da, "sites_original.RDS"))
st_barrier = st_read(file.path(DIR$da, "2020-08-22_all_barriers.gpkg"), quiet = T)

# carpeting ---------------------------------------------------------------
if (st_crs(st_barrier) != st_crs(st_as_sf(dt_rivers))) {
        st_barrier %<>% st_transform(crs = st_crs(st_as_sf(dt_rivers)))
}

# DELETE ---------------------------------------------------------
print("### --- DELETING SEGMENTS --- ###")
dt_rivers  <-
        dt_rivers[!ecoserv_id %in% c(
                
# ...rlp ---------------------------------------------------------------------

                "rlp_55",
                "rlp_56",
                "rlp_203",
                "rlp_212",
                "rlp_214",
                "rlp_374",
                "rlp_248",
                "rlp_740",
                "rlp_787",
                "rlp_1370",
                "rlp_8043",
                "rlp_10137",
                "rlp_10152",
                "rlp_10174",
                "rlp_10420",
                "rlp_10595",
                "rlp_10596",
                "rlp_10597",
                "rlp_10598",
                "rlp_10599",
                "rlp_10604",
                "rlp_10605",
                "rlp_10606",
                "rlp_10610",
                "rlp_10648",
                "rlp_10649",
                "rlp_10650",
                "rlp_10651",
                "rlp_10652",
                "rlp_10653",
                "rlp_10654",
                "rlp_10736",
                "rlp_10737",
                "rlp_10782",
                "rlp_10783",
                "rlp_10820",
                "rlp_10821",
                "rlp_10902",
                "rlp_10918",
                "rlp_10919",
                "rlp_10930",
                "rlp_10931",
                "rlp_10932",
                "rlp_10933",
                "rlp_10941",
                "rlp_10942",
                "rlp_10943",
                "rlp_10944",
                "rlp_10945",
                "rlp_10947",
                "rlp_10948",
                "rlp_10950",
                "rlp_10951",
                "rlp_10965",
                "rlp_10966",
                "rlp_10967",
                "rlp_10968",
                "rlp_10969",
                "rlp_10979",
                "rlp_10981",
                "rlp_10982",
                "rlp_10983",
                "rlp_10988",
                "rlp_10989",
                "rlp_10990",
                "rlp_10991",
                "rlp_10992",
                "rlp_10993",
                "rlp_14271",
                "rlp_16929",
                "rlp_16930",

# ...swd ---------------------------------------------------------------------
                
                "swd_2300",
                "swd_5129",
                "swd_5419",
                "swd_6248",
                "swd_7267",
                "swd_7359",
                "swd_7441",
                "swd_8426",
                "swd_10074",
                "swd_11480",
                "swd_11512",
                "swd_13087",
                "swd_16359",
                "swd_16871",
                "swd_17886",
                "swd_18503",
                "swd_19631",
                "swd_19688",

                "swd_31780",
                "swd_31787",
                "swd_31782",
                "swd_31864",
                "swd_31923",
                "swd_31943",
                "swd_31989",

                "swd_32041",
                "swd_32142",
                "swd_32205",
                "swd_32259",
                "swd_32360",
                "swd_32363",
                "swd_32406",
                "swd_32717",
                "swd_32832",

                "swd_33063",
                "swd_33102",
                "swd_33118",
                "swd_33221",
                "swd_33230",
                "swd_33300",
                "swd_33338",
                "swd_33440",
                "swd_33483",
                "swd_33560",
                "swd_33588",
                "swd_33755",
                "swd_33758",
                "swd_33772",
                "swd_33797",
                "swd_33801",
                "swd_33927",
                "swd_33981",

                "swd_34009",
                "swd_34024",
                "swd_34037",
                "swd_34111",
                "swd_34134",
                "swd_34174",
                "swd_34225",
                "swd_34337",
                "swd_34354",
                "swd_34538",
                "swd_34555",
                "swd_34671",
                "swd_34679",
                "swd_34810",
                "swd_34819",
                "swd_34896",

                "swd_35010",
                "swd_35038",
                "swd_35042",
                "swd_35046",
                "swd_35288",
                "swd_35318",
                "swd_35357",
                "swd_35425",
                "swd_35681",
                "swd_35713",
                "swd_35910",
                "swd_35927",
                "swd_35980",

                "swd_36019",
                "swd_36080",
                "swd_36021",
                "swd_36160",
                "swd_36244",
                "swd_36404",
                "swd_36644",
                "swd_36675",
                "swd_36926",

                "swd_37041",
                "swd_37050",
                "swd_37075",
                "swd_37142",
                "swd_37206",
                "swd_37222",
                "swd_37240",
                "swd_37325",
                "swd_37497",
                "swd_37658",
                "swd_37366",
                "swd_37712",
                "swd_37838",

                "swd_38027",
                "swd_38051",
                "swd_38065",
                "swd_38076",
                "swd_38203",
                "swd_38376",
                "swd_38463",
                "swd_38662",
                "swd_38678",
                "swd_38681",
                "swd_38684",
                "swd_38780",
                "swd_38834",
                "swd_38904",

                "swd_39022",
                "swd_39139",
                "swd_39153",
                "swd_39213",
                "swd_39272",
                "swd_39350",
                "swd_39444",
                "swd_39586",
                "swd_39618",
                "swd_39690",
                "swd_39720",
                "swd_39821",
                "swd_39970",
                "swd_39976",

                "swd_40052",
                "swd_40069",
                "swd_40136",
                "swd_40177",
                "swd_40232",
                "swd_40413",
                "swd_40450",
                "swd_40599",
                "swd_40783",
                "swd_40824",
                "swd_40833",
                "swd_40883",

                "swd_41027",
                "swd_41336",
                "swd_41470",
                "swd_41476",
                "swd_41490",
                "swd_41717",
                "swd_41759",
                "swd_41916",
                "swd_41952",
                "swd_42015",
                "swd_42020",
                "swd_42205",
                "swd_42326",
                "swd_42332",
                "swd_42407",
                "swd_42819",
                "swd_43506",
                "swd_42606",
                "swd_42611",
                "swd_42776",
                "swd_42781",
                "swd_43032",
                "swd_43140",
                "swd_43169",
                "swd_43200",
                "swd_43456",
                "swd_43543",
                "swd_43603",
                "swd_43694",
                "swd_43784",
                "swd_43799",
                "swd_43863",
                "swd_43910",
                "swd_43930",
                "swd_43986",
                "swd_44248",
                "swd_44269",
                "swd_44335",
                "swd_44405",
                "swd_44654",
                "swd_44665",
                "swd_44667",
                "swd_44478",
                "swd_44644",
                "swd_44793",
                "swd_44847",
                "swd_44849",
                "swd_44885",
                "swd_44990",
                "swd_45083",
                "swd_45123",
                "swd_45278",
                "swd_45290",
                "swd_45416",
                "swd_45584",
                "swd_46068",

                "swd_47115",
                "swd_47116",
                "swd_47175",

                "swd_48147",
                "swd_48171",
                "swd_48172",

                "swd_49283",
                "swd_50486",
                "swd_51423",
                "swd_53188",
                "swd_53467",
                "swd_53470",
                "swd_53736",
                "swd_53752",
                "swd_53948",
                "swd_53949",
                "swd_54166",
                "swd_54752",
                "swd_55797",
                "swd_55798",
                "swd_56269",
                "swd_56269",
                "swd_56613",
                "swd_57699",
                "swd_57700",
                "swd_57851",
                "swd_58021",
                "swd_58143",
                "swd_58895",
                "swd_58896",
                "swd_59160",
                "swd_59794",
                "swd_60407",
                "swd_60820",
                "swd_62213",
                "swd_62219",
                "swd_62290",
                "swd_62318",
                "swd_62438",
                "swd_63068",
                "swd_63297",
                "swd_63641",
                "swd_63811",
                "swd_63899",
                "swd_64096",
                "swd_64361",
                "swd_64976",
                "swd_64977",
                "swd_64978",
                "swd_65047",
                "swd_65502",
                "swd_65805",
                "swd_66478",
                "swd_66686",
                "swd_67020",
                "swd_67029",
                "swd_67705",
                "swd_63726",
                "swd_64279",
                "swd_64280",
                "swd_64604",
                "swd_65992",
                "swd_67704",
                "swd_67903",
                "swd_69561",
                "swd_69562",
                "swd_69563",
                "swd_69582",

                "swd_71575",
                "swd_71632",
                "swd_71702",
                "swd_71703",
                "swd_71961",
                "swd_71962",
                "swd_71963",

                "swd_72024",
                "swd_72025",
                "swd_72026",
                "swd_72042",
                "swd_72043",
                "swd_72044",
                "swd_72286",
                "swd_72301",
                "swd_72302",
                "swd_72303",
                "swd_72304",
                "swd_72305",
                "swd_72365",
                "swd_72366",
                "swd_72911",
                "swd_72367",
                "swd_72890",
                "swd_72891",
                "swd_72892",

# ...vdn ---------------------------------------------------------------------
                "vdn_2342",
                "vdn_2791",
                "vdn_3493",
                "vdn_3502",
                "vdn_3553",
                "vdn_3658",
                "vdn_3658",
                "vdn_3658",
                "vdn_3890",
                "vdn_3946",
                "vdn_3946",
                "vdn_5868",
                "vdn_5685",
                "vdn_5687",
                "vdn_5694",
                "vdn_5695",
                "vdn_5696",
                "vdn_5697",
                "vdn_5698",
                "vdn_5820",
                "vdn_5869",
                "vdn_5908",

                "vdn_6655",
                "vdn_6656",
                "vdn_6657",
                "vdn_6658",
                "vdn_6659",
                "vdn_6660",
                "vdn_6661",
                "vdn_6662",
                "vdn_6663",
                "vdn_6664",
                "vdn_6665",
                "vdn_6666",
                "vdn_6667",
                "vdn_6667",
                "vdn_6668",
                "vdn_6669",
                "vdn_6670",
                "vdn_6671",
                "vdn_6672",
                "vdn_6673",
                "vdn_7346",
                "vdn_7763",
                "vdn_7770",
                "vdn_7861",
                "vdn_7985",
                "vdn_8111",
                "vdn_8112",

                "vdn_678"
        )]


# MANUAL -----------------------------------------------------
print("### --- MANUAL IMPROVEMENTS --- ###")
dt_rivers[ecoserv_id == "rlp_54"   , FROM := "P52"]
dt_rivers[ecoserv_id == "rlp_192"  , FROM := "p_add_m_01"]
dt_rivers[ecoserv_id == "rlp_192"  , TO := "p_add_01"]
dt_rivers[ecoserv_id == "rlp_10118", FROM := "P2590"]
dt_rivers[ecoserv_id == "rlp_10143", FROM := "P2634"]
dt_rivers[ecoserv_id == "rlp_10143", TO   := "P2638"]
dt_rivers[ecoserv_id == "rlp_10280", FROM := "P3136"]
dt_rivers[ecoserv_id == "rlp_10389", TO := "Px3136"]
dt_rivers[ecoserv_id == "rlp_10390", FROM := "Px3136"]
dt_rivers[ecoserv_id == "rlp_10390", TO := "P3136"]
dt_rivers[ecoserv_id == "rlp_10391", FROM := "P3136"]
dt_rivers[ecoserv_id == "rlp_10394", FROM := "P2917"]
# dt_rivers[ecoserv_id == "rlp_10398", FROM := "P3150"]
# dt_rivers[ecoserv_id == "rlp_10398", TO   := "P3160"]
dt_rivers[ecoserv_id == "rlp_10398", FROM := "P3152"]
dt_rivers[ecoserv_id == "rlp_10398", TO := "P3154"]
dt_rivers[ecoserv_id == "rlp_10399", FROM := "P3150"]
dt_rivers[ecoserv_id == "rlp_10509", FROM := "P_add_3374"]
dt_rivers[ecoserv_id == "rlp_10560", FROM := "P3474"]
dt_rivers[ecoserv_id == "rlp_10580", FROM := "P3474"]
dt_rivers[ecoserv_id == "rlp_10581", FROM := "P3480"]
dt_rivers[ecoserv_id == "rlp_10589", FROM := "P3512"]
dt_rivers[ecoserv_id == "rlp_10611", FROM := "P3414"]
dt_rivers[ecoserv_id == "rlp_10627", FROM := "P3598"]
dt_rivers[ecoserv_id == "rlp_10627", TO   := "P3612"]
dt_rivers[ecoserv_id == "rlp_10634", FROM := "P3422"]
dt_rivers[ecoserv_id == "rlp_10698", FROM := "P3754"]
dt_rivers[ecoserv_id == "rlp_10698", TO := "P3750"]
dt_rivers[ecoserv_id == "rlp_10717", FROM := "P3792"]
dt_rivers[ecoserv_id == "rlp_10717", TO := "P3760"]
dt_rivers[ecoserv_id == "rlp_10770", FROM := "P288"]

dt_rivers[ecoserv_id == "rlp_14111", FROM := "P29614"]
dt_rivers[ecoserv_id == "rlp_16937", FROM   := "P2917"]
dt_rivers[ecoserv_id == "rlp_16941", FROM := "P3749"]
dt_rivers[ecoserv_id == "rlp_16941", TO := "P3649"]
dt_rivers[ecoserv_id == "rlp_16945", FROM := "P3796"]
dt_rivers[ecoserv_id == "rlp_16945", TO := "P3798"]

dt_rivers[ecoserv_id == "vdn_269" , TO   := "P16648"]
dt_rivers[ecoserv_id == "vdn_973" , TO   := "P29028"]
dt_rivers[ecoserv_id == "vdn_2111" , TO   := "P25900"]
dt_rivers[ecoserv_id == "vdn_2342" , TO   := "P28346"]
dt_rivers[ecoserv_id == "vdn_3872" , TO := "P21634_add"]
dt_rivers[ecoserv_id == "vdn_4415" , FROM := "P21634_add"]
dt_rivers[ecoserv_id == "vdn_2247" , TO := "P26094"]
dt_rivers[ecoserv_id == "vdn_2342" , FROM := "P22684"]
dt_rivers[ecoserv_id == "vdn_3502", TO := "P20926"]
dt_rivers[ecoserv_id == "vdn_4399" , FROM := "P22684"]
dt_rivers[ecoserv_id == "vdn_7362" , FROM := "P20200"]
dt_rivers[ecoserv_id == "vdn_7362" , TO   := "P22508"]
dt_rivers[ecoserv_id == "vdn_7765", FROM := "P17078"]

dt_rivers[ecoserv_id == "vdn_8001" , FROM   := "P21820"]
dt_rivers[ecoserv_id == "vdn_8001" , TO   := "P29426"]

dt_rivers[ecoserv_id == "sar_4600",  FROM := "P10674"]
dt_rivers[ecoserv_id == "sar_8158",  FROM := "P117261"]
dt_rivers[ecoserv_id == "sar_8158",  TO   := "P114079"]
dt_rivers[ecoserv_id == "sar_8400",  FROM := "P122813"]
dt_rivers[ecoserv_id == "sar_8400",  TO   := "P403"]
dt_rivers[ecoserv_id == "vdn_16960", FROM := "P10564"]
dt_rivers[ecoserv_id == "swd_45552", TO := "P113424"]
dt_rivers[ecoserv_id == "swd_2954", TO := "P31297"]

dt_rivers[ecoserv_id == "add_72920", FROM := "P3164"]

# REVERSE -----------------------------------------------------------------
print("### --- REVERSE SEGMENTS --- ###")
dt_rivers = reverse(x= c(

# ... rlp -----------------------------------------------------------------
        "rlp_197", 
        "rlp_537",
        "rlp_3996",
        "rlp_5003",
        "rlp_6082",
        "rlp_7990",
        "rlp_10080",
        "rlp_10081",
        "rlp_10107",
        "rlp_10108",
        "rlp_10109",
        "rlp_10116",
        "rlp_10117",
        "rlp_10138",
        "rlp_10139",
        "rlp_10141",
        "rlp_10145",
        "rlp_10146",
        "rlp_10279",
        "rlp_10283",
        "rlp_10284",
        "rlp_10285",
        "rlp_10333", 
        "rlp_10334",
        "rlp_10338",
        "rlp_10400",
        "rlp_10406", 
        "rlp_10407",
        "rlp_10418",
        "rlp_10470",
        "rlp_10478",
        "rlp_10506",
        "rlp_10557",
        "rlp_10558",
        "rlp_10559",
        "rlp_10562",
        "rlp_10584",
        "rlp_10585",
        "rlp_10592",
        "rlp_10593",
        "rlp_10632",
        "rlp_10639",
        "rlp_10640",
        "rlp_10642",
        "rlp_10645",
        "rlp_10673",
        "rlp_10725",
        "rlp_10726",
        "rlp_10727",
        "rlp_10768",
        "rlp_10770",
        "rlp_10771",
        "rlp_10774",
        "rlp_11002",
        "rlp_11005",
        "rlp_11007",
        "rlp_11011",
        "rlp_11015",
        "rlp_11019",
        "rlp_11021",
        "rlp_11132",
        "rlp_11140",
        "rlp_11158",
        "rlp_11159",
        "rlp_11160",
        "rlp_11177",
        "rlp_11199",

        "rlp_11219",

        "rlp_11220",
        "rlp_11221",
        "rlp_11273",
        "rlp_11298",
        "rlp_11309",
        "rlp_11315",
        "rlp_11331",
        "rlp_11401",
        "rlp_11448",
        
        "rlp_11718",
        "rlp_11758",
        "rlp_12282",
        "rlp_12283",
        "rlp_12737",
        "rlp_12748",
        "rlp_12784",
        "rlp_12788",
        "rlp_12800",
        "rlp_12822",
        "rlp_12836",
        "rlp_12862",
        "rlp_12872",
        "rlp_12874",
        "rlp_12875",
        "rlp_12876",
        "rlp_12908",
        "rlp_12917",
        "rlp_13302",
        "rlp_13310",
        "rlp_13313",
        "rlp_13321",
        "rlp_13331",
        "rlp_13337",
        "rlp_13349",
        "rlp_13352",
        "rlp_13362",
        "rlp_13386",
        "rlp_13395",
        "rlp_13396",
        "rlp_13397",
        "rlp_13400",
        "rlp_13406",
        "rlp_13409",
        "rlp_13444",
        "rlp_13454",
        "rlp_13459",
        "rlp_13461",
        "rlp_13462",
        "rlp_13463",
        "rlp_13465",
        "rlp_13468",
        "rlp_13469",
        "rlp_13471",
        "rlp_13481",
        "rlp_13482",
        "rlp_13503",
        "rlp_13507",
        "rlp_13511",
        "rlp_13516",
        "rlp_13517",
        "rlp_13541",
        "rlp_13608",
        "rlp_13627",
        "rlp_13640",
        "rlp_13659",
        "rlp_13661",
        "rlp_13663",
        "rlp_13705",
        "rlp_13711",
        "rlp_13718",
        "rlp_13752",
        "rlp_13757",
        "rlp_13762",
        "rlp_13823",
        "rlp_13843",
        "rlp_13846",
        "rlp_13889",
        "rlp_13909",
        "rlp_13912",
        "rlp_13916",
        "rlp_13927",
        "rlp_13961",
        "rlp_13974",
        "rlp_13977",
        "rlp_14037",
        "rlp_14045",
        "rlp_14100",
        "rlp_14146",
        "rlp_14156",
        "rlp_14258",
        "rlp_14291",
        "rlp_14301",
        "rlp_14302",
        "rlp_14423",
        "rlp_14473",
        "rlp_14488",
        "rlp_14513",
        "rlp_14532",
        "rlp_14536",
        "rlp_14538",
        "rlp_14561",
        "rlp_14587",
        "rlp_14608",
        "rlp_14622",
        "rlp_14639",
        "rlp_14644",
        "rlp_14658",
        "rlp_14659",
        "rlp_14665",
        "rlp_14696",
        "rlp_14697",
        #"rlp_14718",
        "rlp_14719",
        "rlp_14750",
        "rlp_14765",
        "rlp_14778",
        "rlp_14832",
        "rlp_14869",

        "rlp_14923",
        "rlp_14932",
        "rlp_14934",
        "rlp_14951",
        
        "rlp_15004",
        "rlp_15012",
        "rlp_15056",
        "rlp_15082",
        "rlp_15096",
        "rlp_15120",
        "rlp_15187",
        "rlp_15191",
        "rlp_15198",
        "rlp_15253",
        "rlp_15265",
        "rlp_15284",
        "rlp_15286",
        "rlp_15335",
        "rlp_15378",
        "rlp_15382",
        "rlp_15389",
        "rlp_15410",
        "rlp_15434",
        "rlp_15550",
        "rlp_15599",
        "rlp_15636",
        "rlp_15690",
        "rlp_15703",
        "rlp_15760",
        "rlp_15765",
        "rlp_15804",
        "rlp_15808",
        "rlp_15831",
        "rlp_15905",
        "rlp_16348",
        "rlp_16613",
        "rlp_16614",
        "rlp_16615",
        "rlp_16623",
        "rlp_16626",
        "rlp_16954",
        "rlp_16955",
        "rlp_16956",

# ... vdn -----------------------------------------------------------------

        
        "vdn_85",
        #"vdn_247",
        "vdn_297",
        "vdn_425",
        "vdn_507",
        "vdn_758",
        "vdn_900",
        "vdn_1011",
        "vdn_1209",
        "vdn_1328",
        "vdn_1381",
        "vdn_1760",
        "vdn_1928",
        #"vdn_2088",
        "vdn_2275",
        "vdn_2379",
        "vdn_2513",
        "vdn_2571",
        "vdn_2796",
        "vdn_3048",
        "vdn_3240",
        "vdn_3489",
        "vdn_3493",
        "vdn_3494",
        "vdn_3502",
        "vdn_3505",
        "vdn_3507",
        "vdn_3508",
        "vdn_3522",
        "vdn_3542",
        "vdn_3587",
        "vdn_3588",
        "vdn_3589",
        "vdn_3652",
        "vdn_3667",
        "vdn_3669",
        "vdn_3686",
        "vdn_3694",
        "vdn_3697",
        "vdn_3699",
        "vdn_3754",
        "vdn_3794",
        "vdn_3804",
        "vdn_3871",
        "vdn_3875",
        "vdn_3890",
        "vdn_3909",
        "vdn_3935",
        "vdn_3961",
        "vdn_4090",
        "vdn_4102",
        "vdn_4178",
        "vdn_4246",
        "vdn_4256",
        "vdn_4265",
        "vdn_4292",
        "vdn_4321",
        "vdn_4349",
        "vdn_4370",
        "vdn_4371",
        "vdn_4373",
        "vdn_4382",
        "vdn_4391",
        "vdn_4394",
        "vdn_4395",
        "vdn_4430",
        "vdn_4588",
        "vdn_4595",
        "vdn_4605",
        "vdn_4628",
        "vdn_4633",
        "vdn_4722",
        "vdn_4826",
        "vdn_4838",
        "vdn_4856",
        "vdn_4863",
        "vdn_4878",
        "vdn_4974",
        "vdn_4976",
        "vdn_5034",
        "vdn_5039",
        "vdn_5044",
        "vdn_5051",
        "vdn_5099",
        "vdn_5160",
        "vdn_5163",
        "vdn_5199",
        "vdn_5245",
        "vdn_5258",
        "vdn_5279",
        "vdn_5280",
        "vdn_5361",
        "vdn_5367",
        "vdn_5368",
        "vdn_5382",
        "vdn_5403",
        "vdn_5420",
        "vdn_5421",
        "vdn_5427",
        "vdn_5440",
        "vdn_5448",
        "vdn_5449",
        "vdn_5465",
        "vdn_5467",
        "vdn_5471",
        "vdn_5622",
        "vdn_5626",
        "vdn_5717",
        "vdn_5750",
        "vdn_5812",
        "vdn_5826",
        "vdn_5883",
        "vdn_5887",
        "vdn_5890",
        "vdn_5901",
        "vdn_5902",
        "vdn_5904",
        "vdn_5925",
        "vdn_5936",
        "vdn_5946",
        "vdn_6006",
        "vdn_6008",
        "vdn_6027",
        "vdn_6028",
        "vdn_6051",
        "vdn_6085",
        "vdn_6093",
        "vdn_6103",
        "vdn_6120",
        "vdn_6121",
        "vdn_6128",
        "vdn_6159",
        "vdn_6161",
        "vdn_6164",
        "vdn_6192",
        "vdn_6193",
        "vdn_6196",
        "vdn_6256",
        "vdn_6298",
        "vdn_6337",
        "vdn_6352",
        "vdn_6356",
        "vdn_6365",
        "vdn_6369",
        "vdn_6384",
        "vdn_6387",
        "vdn_6388",
        "vdn_6393",
        "vdn_6395",
        "vdn_6397",
        "vdn_6400",
        "vdn_6401",
        "vdn_6402",
        "vdn_6407",
        "vdn_6410",
        "vdn_6413",
        "vdn_6415",
        "vdn_6417",
        "vdn_6422",
        "vdn_6423",
        "vdn_6426",
        "vdn_6431",
        "vdn_6439",
        "vdn_6442",
        "vdn_6445",
        "vdn_6450",
        "vdn_6452",
        "vdn_6455",
        "vdn_6457",
        "vdn_6461",
        "vdn_6463",
        "vdn_6465",
        "vdn_6468",
        "vdn_6471",
        "vdn_6473",
        "vdn_6476",
        "vdn_6477",
        "vdn_6479",
        "vdn_6482",
        "vdn_6483",
        "vdn_6485",
        "vdn_6486",
        "vdn_6492",
        "vdn_6497",
        "vdn_6499",
        "vdn_6502",
        "vdn_6504",
        "vdn_6505",
        "vdn_6507",
        "vdn_6508",
        "vdn_6509",
        "vdn_6510",
        "vdn_6512",
        "vdn_6515",
        "vdn_6517",
        "vdn_6520",
        "vdn_6522",
        "vdn_6523",
        "vdn_6525",
        "vdn_6526",
        "vdn_6528",
        "vdn_6530",
        "vdn_6531",
        "vdn_6534",
        "vdn_6536",
        "vdn_6537",
        "vdn_6538",
        "vdn_6541",
        "vdn_6544",
        "vdn_6545",
        "vdn_6547",
        "vdn_6548",
        "vdn_6553",
        "vdn_6555",
        "vdn_6558",
        "vdn_6565",
        "vdn_6566",
        "vdn_6568",
        "vdn_6570",
        "vdn_6574",
        "vdn_6576",
        "vdn_6577",
        "vdn_6578",
        "vdn_6580",
        "vdn_6581",
        "vdn_6582",
        "vdn_6586",
        "vdn_6590",
        "vdn_6591",
        "vdn_6595",
        "vdn_6598",
        "vdn_6601",
        "vdn_6674",
        "vdn_6677",
        "vdn_6683",
        "vdn_6687",
        "vdn_6700",
        "vdn_6703",
        "vdn_6709",
        "vdn_6713",
        "vdn_6716",
        "vdn_6729",
        "vdn_6731",
        "vdn_6742",
        "vdn_6761",
        "vdn_6770",
        "vdn_6792",
        "vdn_6795",
        "vdn_6797",
        "vdn_6799",
        "vdn_6801",
        "vdn_6804",
        "vdn_6809",
        "vdn_6813",
        "vdn_6823",
        "vdn_6832",
        "vdn_6833",
        "vdn_6834",
        "vdn_6878",
        "vdn_6951",
        "vdn_6952",
        "vdn_6981",
        "vdn_7026",
        "vdn_7028",
        "vdn_7089",
        "vdn_7102",
        "vdn_7107",
        "vdn_7113",
        "vdn_7119",
        "vdn_7120",
        "vdn_7126",
        "vdn_7131",
        "vdn_7134",
        "vdn_7143",
        "vdn_7165",
        "vdn_7199",
        "vdn_7212",
        "vdn_7255",
        "vdn_7262",
        "vdn_7266",
        "vdn_7267",
        "vdn_7268",
        "vdn_7282",
        "vdn_7300",
        "vdn_7304",
        "vdn_7345",
        "vdn_7369",
        "vdn_7371",
        "vdn_7376",
        "vdn_7384",
        "vdn_7386",
        "vdn_7391",
        "vdn_7396",
        "vdn_7406",
        "vdn_7410",
        "vdn_7411",
        
        "vdn_7413",
        "vdn_7424",
        "vdn_7434",
        "vdn_7468",
        #"vdn_7469",
        "vdn_7470",
        "vdn_7571",
        "vdn_7581",
        "vdn_7586",
        "vdn_7588",
        "vdn_7590",
        "vdn_7594",
        "vdn_7596",
        "vdn_7603",
        "vdn_7604",
        "vdn_7608",
        "vdn_7617",
        "vdn_7620",
        "vdn_7622",
        "vdn_7637",
        "vdn_7683",
        
        "vdn_7687",
        "vdn_7696",
        "vdn_7698",
        "vdn_7704",
        "vdn_7705",
        "vdn_7708",
        "vdn_7742",
        "vdn_7743",
        "vdn_7767",
        "vdn_7768",
        "vdn_7772",
        "vdn_7802",
        "vdn_7806",
        "vdn_7809",
        "vdn_7811",
        "vdn_7930",
        "vdn_7933",
        "vdn_7935",
        "vdn_7942",
        "vdn_7953",
        "vdn_7959",
        "vdn_7962",
        #"vdn_7963",
        #"vdn_7964",
        "vdn_7965",
        "vdn_7966",
        "vdn_7967",
        "vdn_7974",
        "vdn_7978",
        "vdn_7980",
        "vdn_7994",
        "vdn_7998",
        "vdn_8010",
        "vdn_8011",
        "vdn_8018",
        "vdn_8039",
        "vdn_8045",
        "vdn_8047",
        "vdn_8065",
        "vdn_8067",
        "vdn_8101",
        "vdn_8102",
        "vdn_8114",
        "vdn_8137",
        "vdn_8139",
        "vdn_8144",
        "vdn_8150",
        "vdn_8153",
        "vdn_8155",
        "vdn_8156",
        "vdn_8158",
        "vdn_8165",
        "vdn_8169",
        "vdn_8176",
        "vdn_8181",
        "vdn_8182",
        "vdn_8185",
        "vdn_8187",
        "vdn_8190",
        "vdn_8191",
        "vdn_8192",
        "vdn_8202",
        "vdn_8232",
        "vdn_8233",
        "vdn_8239",
        "vdn_8244",
        "vdn_8277",
        "vdn_8280",
        "vdn_8293",
        "vdn_8295",
        "vdn_8301",
        "vdn_8315",
        "vdn_8317",
        "vdn_8319",
        

# ... swd ---------------------------------------------------------------------
        "swd_32047",
        "swd_32091",
        "swd_32282",
        "swd_32284",
        "swd_32441",
        "swd_32551",
        "swd_32568",
        "swd_32579",
        "swd_32594",
        "swd_32680",
        "swd_32698",
        "swd_32919",
        "swd_33016",
        "swd_33127",
        "swd_33164",
        "swd_33251",
        "swd_33367",
        "swd_33426",
        "swd_33514",
        "swd_33551",
        "swd_33619",
        "swd_33712",
        "swd_33791",
        "swd_34588",
        "swd_34685",
        "swd_34801",
        "swd_35017",
        "swd_35140",
        "swd_35322",
        "swd_35336",
        "swd_35381",
        "swd_35480",
        "swd_35584",
        "swd_35687",
        "swd_35793",
        "swd_35808",
        "swd_35907",
        "swd_35941",
        "swd_35970",
        "swd_36030",
        "swd_36093",
        "swd_36299",
        "swd_36323",
        "swd_36454",
        "swd_36827",
        "swd_36928",
        "swd_36953",
        "swd_36973",
        "swd_37200",
        "swd_37373",
        "swd_37397",
        "swd_37737",
        "swd_37755",
        "swd_37940",
        "swd_38160",
        "swd_38324",
        "swd_38384",
        "swd_38425",
        "swd_38599",
        "swd_38667",
        "swd_38772",
        "swd_38885",
        "swd_38900",
        "swd_39236",
        "swd_39343",
        "swd_39448",
        "swd_39690",
        "swd_39765",
        "swd_39908",

        "swd_40207",
        "swd_40255",
        "swd_40316",
        "swd_40441",
        "swd_40505",
        "swd_40506",
        "swd_41509",
        "swd_40510",
        "swd_40692",
        "swd_40860",
        "swd_40865",
        "swd_41006",
        "swd_41102",
        "swd_41330",
        "swd_41414",
        "swd_41428",
        "swd_41483",
        "swd_41761",
        "swd_41855",
        "swd_41897",
        "swd_41929",
        "swd_42037",
        "swd_42311",
        "swd_42348",
        "swd_42540",
        "swd_42588",
        "swd_42629",
        "swd_42672",
        "swd_42807",
        "swd_42830",
        "swd_42833",
        "swd_42958",
        "swd_43001",
        "swd_43249",
        "swd_43289",
        "swd_43448",
        "swd_43482",
        "swd_43777",
        "swd_44050",
        "swd_44288",
        "swd_44503",
        "swd_44586",
        "swd_44620",
        "swd_44656",
        "swd_44691",
        "swd_44723",
        "swd_44796",
        "swd_44913",
        "swd_44955",
        "swd_45008",
        "swd_45128",
        "swd_45188",
        "swd_45247",
        "swd_45306",
        "swd_45355",
        "swd_45477",
        "swd_45610",
        "swd_45657",
        "swd_45664",
        "swd_45666",
        "swd_45669",
        "swd_45699",
        "swd_45702",
        "swd_45771",
        "swd_45895",
        "swd_45899",
        "swd_45902",
        "swd_45908",
        "swd_45918",
        "swd_45924",
        "swd_45927",
        "swd_45992",

        "swd_46006",
        "swd_46016",
        "swd_46018",
        "swd_46042",
        "swd_46097",
        "swd_46113",
        "swd_46116",
        "swd_46166",
        "swd_46167",
        "swd_46169",
        "swd_46172",
        "swd_46175",
        "swd_46180",
        "swd_46182",
        "swd_46184",
        "swd_46185",
        "swd_46187",
        "swd_46192",
        "swd_46266",
        "swd_46267",
        "swd_46285",
        "swd_46288",
        "swd_46294",
        "swd_46300",
        "swd_46306",
        "swd_46337",
        "swd_46341",
        "swd_46343",
        "swd_46371",
        "swd_46387",
        "swd_46388",
        "swd_46393",
        "swd_46412",
        "swd_46440",
        "swd_46441",
        "swd_46460",
        "swd_46522",
        "swd_46524",
        "swd_46527",
        "swd_46606",
        "swd_46625",
        "swd_46655",
        "swd_46664",
        "swd_46724",
        "swd_46725",
        "swd_46760",
        "swd_46818",
        "swd_46829",
        "swd_46832",
        "swd_46841",
        "swd_46843",
        "swd_46848",
        "swd_46854",
        "swd_46856",
        "swd_46866",
        "swd_47160",
        "swd_47191",
        "swd_47206",
        "swd_47227",
        "swd_47255",
        "swd_47314",
        "swd_47322",
        "swd_47382",
        "swd_47394",
        "swd_47396",
        "swd_47401",
        "swd_47408",
        "swd_47416",
        "swd_47417",
        "swd_47419",
        "swd_47420",
        "swd_47422",
        "swd_47424",
        "swd_47433",
        "swd_47440",
        "swd_47446",
        "swd_47452",
        "swd_47461",
        "swd_47462",
        "swd_47464",
        "swd_47465",
        "swd_47469",
        "swd_47470",
        "swd_47472",
        "swd_47475",
        "swd_47479",
        "swd_47483",
        "swd_47487",
        "swd_47489",
        "swd_47493",
        "swd_47512",
        "swd_47539",
        "swd_47618",
        "swd_47631",
        "swd_47632",
        "swd_47647",
        "swd_47651",
        "swd_47652",
        "swd_47719",
        "swd_47762",
        "swd_47764",
        "swd_47930",
        "swd_48098",
        "swd_48106",
        "swd_48120",
        "swd_48121",
        "swd_48178",
        "swd_48201",
        "swd_48203",
        "swd_48204",
        "swd_48205",
        "swd_48266",
        "swd_48267",
        "swd_48308",
        "swd_48326",
        "swd_48641",
        "swd_48642",
        "swd_48789",
        "swd_48866",
        "swd_48872",
        "swd_48874",

        "swd_49037",
        "swd_49051",
        "swd_49112",
        "swd_49187",
        "swd_49190",
        "swd_49244",
        "swd_49284",
        "swd_49300",
        "swd_49435",
        "swd_49535",
        "swd_49589",
        "swd_49693",
        "swd_49685",
        "swd_49791",
        "swd_49793",
        "swd_49795",
        "swd_49797",
        "swd_49810",
        "swd_49812",
        "swd_49818",
        "swd_49820",
        "swd_49920",
        "swd_49922",
        "swd_49933",
        "swd_49988",
        "swd_49994",
        "swd_49997",

        
        "swd_50000",
        "swd_50011",
        "swd_50031",
        "swd_50037",
        "swd_50039",
        "swd_50040",
        "swd_50043",
        "swd_50047",
        "swd_50065",
        "swd_50072",
        "swd_50082",
        "swd_50083",
        "swd_50088",
        "swd_59108",
        "swd_50124",
        "swd_50125",
        "swd_50130",
        "swd_50131",
        "swd_50138",
        "swd_50142",
        "swd_50155",
        "swd_50316",
        "swd_50422",
        "swd_50466",
        "swd_50471",
        "swd_50535",
        "swd_50565",
        "swd_50566",
        "swd_50829",
        "swd_50845",
        "swd_51082",
        "swd_51087",
        "swd_51088",
        "swd_51090",
        "swd_51100",
        "swd_51106",
        "swd_51112",
        "swd_51213",
        "swd_51214",
        "swd_51277",
        "swd_51332",
        "swd_51364",
        "swd_51398",
        "swd_51455",
        "swd_51726",
        "swd_51776",
        "swd_51789",
        "swd_51790",
        "swd_51792",
        "swd_51794",
        "swd_51806",
        "swd_51809",
        "swd_51815",
        "swd_51822",
        "swd_51826",
        "swd_51866",
        "swd_51873",
        "swd_52026",
        "swd_52041",
        "swd_52042",
        "swd_52045",
        "swd_52048",
        "swd_52051",
        "swd_52052",
        "swd_52054",
        "swd_52062",
        "swd_52069",
        "swd_52070",
        "swd_52072",
        "swd_52079",
        "swd_52223",
        "swd_52224",
        "swd_52226",
        "swd_52291",
        "swd_52533",
        "swd_52614",
        "swd_52668",
        "swd_52671",
        "swd_52852",
        "swd_52857",
        "swd_52878",
        "swd_52982",
        "swd_51992",
        "swd_52266",
        "swd_52822",
        "swd_52824",
        "swd_52827",
        "swd_52830",
        
        "swd_53224",
        "swd_53244",
        "swd_53466",
        "swd_53471",
        "swd_53517",
        "swd_53519",
        "swd_53541",
        "swd_53542",
        "swd_53543",
        "swd_53544",
        "swd_53545",
        "swd_53546",
        "swd_53547",
        "swd_53548",
        "swd_53549",
        "swd_53550",
        "swd_53551",
        "swd_53552",
        "swd_53553",
        "swd_53557",
        "swd_53746",
        "swd_53771",
        "swd_53873",
        "swd_53945",
        "swd_53984",
        "swd_53997",
        
        "swd_54002",
        "swd_54003",
        "swd_54011",
        "swd_54022",
        "swd_54067",
        "swd_54073",
        "swd_54082",
        "swd_54096",
        "swd_54099",
        "swd_54107",
        "swd_54115",
        "swd_54124",
        "swd_54126",
        "swd_54129",
        "swd_54142",
        "swd_54158",
        "swd_54168",
        "swd_54172",
        "swd_54339",
        "swd_54348",
        "swd_54349",
        "swd_54415",
        "swd_54417",
        "swd_54423",
        "swd_54424",
        "swd_54427",
        "swd_54429",
        "swd_54435",
        "swd_54444",
        "swd_54450",
        "swd_54467",
        "swd_54481",
        "swd_54514",
        "swd_54587",
        "swd_54602",
        "swd_54649",
        "swd_54707",
        "swd_54711",
        "swd_54749",
        "swd_54812",
        "swd_54845",
        "swd_54880",
        "swd_54882",
        "swd_54883",
        "swd_54884",
        "swd_54888",
        "swd_54891",
        "swd_54911",
        
        "swd_55051",
        "swd_55070",
        "swd_55074",
        "swd_55075",
        "swd_55077",
        "swd_55079",
        "swd_55172",
        "swd_55187",
        "swd_55188",
        "swd_55183",
        "swd_55349",
        "swd_55365",
        "swd_55370",
        "swd_55706",
        "swd_55716",
        "swd_55749",
        "swd_55779",
        "swd_55780",
        "swd_55785",
        "swd_55807",
        "swd_55816",
        "swd_55846",
        "swd_55847",
        "swd_55848",
        "swd_55971",
        "swd_55979",
        "swd_55981",
        "swd_55987",
        "swd_55991",

        "swd_56071",
        "swd_56193",
        "swd_56199",
        "swd_56313",
        "swd_56317",
        "swd_56318",
        "swd_56330",
        "swd_56389",
        "swd_56513",
        "swd_56521",
        "swd_56523",
        "swd_56525",
        "swd_56534",
        "swd_56535",
        "swd_56536",
        "swd_56538",
        "swd_56545",
        "swd_56550",
        "swd_56551",
        "swd_56609",
        "swd_56612",
        "swd_56636",
        "swd_56654",
        "swd_56666",
        "swd_56697",
        "swd_56704",
        "swd_56708",
        "swd_56720",
        "swd_56728",
        "swd_56864",
        "swd_56906",


        "swd_57382",
        "swd_57385",
        "swd_57508",
        "swd_57509",
        "swd_57511",
        "swd_57513",
        "swd_57514",
        "swd_57551",
        "swd_57618",
        "swd_57655",
        "swd_57690",
        "swd_57750",
        "swd_57754",
        "swd_57877",
        "swd_57940",
        "swd_57957",
        "swd_58092",
        "swd_58093",
        "swd_58094",
        "swd_58112",
        "swd_58157",
        "swd_58164",
        "swd_58165",
        "swd_58335",
        "swd_58468",
        "swd_58541",
        "swd_58542",
        "swd_58683",
        "swd_58688",
        "swd_58732",
        "swd_58767",
        "swd_58801",
        "swd_58804",
        "swd_58852",
        "swd_58902",
        "swd_58905",
        "swd_58906",
        "swd_59397",
        "swd_59398",

        "swd_59045",
        "swd_59064",
        "swd_59065",
        "swd_59067",
        "swd_59146",
        "swd_59069",
        "swd_59167",
        "swd_59207",
        "swd_59209",
        "swd_59212",
        "swd_59470",
        "swd_59472",
        "swd_59075",
        "swd_59077",
        "swd_59081",
        "swd_59227",
        "swd_59233",
        "swd_59239",
        "swd_59241",
        "swd_59245",
        "swd_59401",
        "swd_59433",
        "swd_59616",
        "swd_59638",
        "swd_59670",
        "swd_59686",
        "swd_59956",
        "swd_59963",
        "swd_59964",

        "swd_60028",
        "swd_60058",
        "swd_60061",
        "swd_60081",
        "swd_60101",
        "swd_60158",
        "swd_60187",
        "swd_60202",
        "swd_60207",
        "swd_60247",
        "swd_60313",
        "swd_60314",
        "swd_60318",
        "swd_60324",
        "swd_60327",
        "swd_60329",
        "swd_60339",
        "swd_60363",
        "swd_60398",
        "swd_60402",
        "swd_60405",
        "swd_60434",
        "swd_60435",
        "swd_60436",
        "swd_60440",
        "swd_60469",
        "swd_60482",
        "swd_60508",
        "swd_60509",
        "swd_60521",
        "swd_60707",
        "swd_60710",
        "swd_60725",
        "swd_60811",
        "swd_60814",
        "swd_60942",

        "swd_61010",
        "swd_61027",
        "swd_61033",
        "swd_61039",
        "swd_61061",
        "swd_61063",
        "swd_61067",
        "swd_61068",
        "swd_61069",
        "swd_61072",
        "swd_61080",
        "swd_61082",
        "swd_61083",
        "swd_61085",
        "swd_61092",
        "swd_61096",
        "swd_61099",
        
        "swd_61112",
        "swd_61235",
        "swd_61261",
        "swd_61262",
        "swd_61384",
        "swd_61427",
        "swd_61561",
        "swd_61589",
        "swd_61638",
        "swd_61639",
        "swd_61642",
        "swd_61643",
        "swd_61646",
        "swd_61657",
        "swd_61660",
        "swd_61675",
        "swd_61727",
        "swd_61817",
        "swd_61818",
        "swd_61827",
        "swd_61828",
        "swd_61831",
        "swd_61833",
        "swd_61834",
        "swd_61842",
        "swd_61844",
        "swd_61847",
        "swd_61848",
        "swd_61850",
        "swd_61853",
        "swd_61854",
        "swd_61857",
        "swd_61858",
        "swd_61859",
        "swd_61861",
        "swd_61865",
        "swd_61866",
        "swd_61867",
        "swd_61868",
        "swd_61870",
        "swd_61873",
        "swd_61877",
        "swd_61881",
        "swd_61883",

        "swd_62033",
        "swd_62122",
        "swd_62132",
        "swd_62133",
        "swd_62229",
        "swd_62239",
        "swd_62241",
        "swd_62272",
        "swd_62322",
        "swd_62380",
        "swd_62431",
        "swd_62433",
        "swd_62562",
        "swd_62628",
        "swd_62662",
        "swd_62686",
        "swd_62760",
        "swd_62761",

        "swd_63050",
        "swd_63064",
        "swd_63067",
        "swd_63069",
        "swd_63070",
        "swd_63077",
        "swd_63078",
        "swd_63083",
        "swd_63085",
        "swd_63124",
        "swd_63135",
        "swd_63136",
        "swd_63164",
        "swd_63165",
        "swd_63166",
        "swd_64217",
        "swd_63220",
        "swd_63240",
        "swd_63243",
        "swd_63259",
        "swd_63281",
        "swd_63311",
        "swd_63324",
        "swd_63327",
        "swd_63337",
        "swd_63344",
        "swd_63355",
        "swd_63364",
        "swd_63365",
        "swd_63431",
        "swd_63435",
        "swd_63438",
        "swd_63441",
        "swd_63442",
        "swd_63478",
        "swd_63486",
        "swd_63490",
        "swd_63495",
        "swd_63499",
        "swd_63501",
        "swd_63528",
        "swd_63532",
        "swd_63533",
        "swd_63545",
        "swd_63562",
        "swd_63565",
        "swd_63588",
        "swd_63620",
        "swd_63629",
        "swd_63643",
        "swd_64730",
        "swd_63779",
        "swd_63794",
        "swd_63795",
        "swd_64825",
        "swd_63864",
        "swd_63901",
        "swd_63949",
        "swd_63956",
        "swd_63971",
        "swd_63983",

        "swd_64002",
        "swd_64039",
        "swd_64082",
        "swd_64290",
        "swd_64317",
        "swd_64337",
        "swd_64431",
        "swd_64453",
        "swd_64456",
        "swd_64460",
        "swd_64463",
        "swd_64465",
        "swd_64478",
        "swd_64582",
        "swd_64586",
        "swd_64589",
        "swd_64838",
        "swd_64841",
        "swd_64843",
        "swd_64844",
        "swd_64902",
        "swd_64927",
        "swd_64929",
        "swd_64933",

        "swd_65027",
        "swd_65028",
        "swd_65029",
        "swd_65204",
        "swd_65206",
        "swd_65287",
        "swd_65293",
        "swd_65356",
        "swd_65393",
        "swd_65400",
        "swd_65600",
        "swd_65644",
        "swd_65663",
        "swd_65684",
        "swd_65746",
        "swd_65783",
        "swd_65811",
        "swd_65850",
        "swd_65855",
        "swd_65858",
        "swd_65859",
        "swd_65861",
        "swd_65864",
        "swd_65867",
        "swd_65868",
        "swd_65873",
        "swd_65874",
        "swd_65876",
        "swd_65877",
        "swd_65879",
        "swd_65882",
        "swd_65886",
        "swd_65887",
        "swd_65888",
        "swd_65889",
        "swd_65891",
        "swd_65894",
        "swd_65895",
        "swd_65899",
        "swd_65904",
        "swd_65906",
        "swd_65908",
        "swd_65917",
        "swd_65920",
        "swd_65922",
        "swd_65923",
        "swd_65926",
        "swd_65929",
        "swd_65931",
        "swd_65933",
        "swd_65936",
        "swd_65994",

        "swd_66166",
        "swd_66216",
        "swd_66231",
        "swd_66299",
        "swd_66301",
        "swd_66303",
        "swd_66305",
        "swd_66312",
        "swd_66316",
        "swd_66319",
        "swd_66342",
        "swd_66353",
        "swd_66360",
        "swd_66434",
        "swd_66436",
        "swd_66439",
        "swd_66461",
        "swd_66464",
        "swd_66514",
        "swd_66518",
        "swd_66587",
        "swd_66588",
        "swd_66590",
        "swd_66608",
        "swd_66647",
        "swd_66672",
        "swd_66694",
        "swd_66700",
        "swd_66715",
        "swd_66767",
        "swd_66792",
        "swd_66823",
        "swd_66866",
        "swd_66884",
        "swd_66892",
        "swd_66988",
        "swd_66994",
        "swd_67004",
        "swd_67031",
        "swd_67033",
        "swd_67040",
        "swd_67053",
        "swd_67056",
        "swd_67103",
        "swd_67135",
        "swd_67148",
        "swd_67238",
        "swd_67357",
        "swd_67409",
        "swd_67410",
        "swd_67424",
        "swd_67426",
        "swd_67451",
        "swd_67490",
        "swd_67492",
        "swd_67585",
        "swd_67559",
        "swd_67659",
        "swd_67681",
        "swd_67683",
        "swd_67686",
        "swd_67687",
        "swd_67688",
        "swd_67690",
        "swd_67692",
        "swd_67694",
        "swd_67845",
        "swd_67957",
        "swd_67972",
        "swd_67989",
        "swd_67990",

        "swd_68007",
        "swd_68114",
        "swd_68110",
        "swd_68105",
        "swd_68104",
        "swd_68107",
        "swd_68109",
        "swd_68143",
        "swd_68145",
        "swd_68146",
        "swd_68147",
        "swd_68155",
        "swd_68157", 
        "swd_68181",
        "swd_68203",
        "swd_68218",
        "swd_68216",
        "swd_68233",
        "swd_68258",
        "swd_68266",
        "swd_68267",
        "swd_68268",
        "swd_68280",
        "swd_68298",
        "swd_68322",
        "swd_68323",
        "swd_68324",
        "swd_68325",
        "swd_68326",
        "swd_68327",
        "swd_68328",
        "swd_68329",
        "swd_68331",
        "swd_68353",
        "swd_68356",
        "swd_68357",
        "swd_68359",
        "swd_68360",
        "swd_68361",
        "swd_68362",
        "swd_68364",
        "swd_68365",
        "swd_68366",
        "swd_68367",
        "swd_68368",
        "swd_68371",
        "swd_68372",
        "swd_68374",
        "swd_68381",
        "swd_68382",
        "swd_68383",
        "swd_68384",
        "swd_68387",
        "swd_68388",
        "swd_68563",
        "swd_68566",
        "swd_68573",
        "swd_68578",
        "swd_68580",
        "swd_68584",
        "swd_68589",
        "swd_68590",
        "swd_68608",
        "swd_68611",
        "swd_68699",
        "swd_68706",
        "swd_68712",
        "swd_68788",
        "swd_68789",
        "swd_68810",
        "swd_68839",
        "swd_68991",
        "swd_68992",
        
        "swd_69120",
        "swd_69178",
        "swd_69180",
        "swd_69230",
        "swd_69233",
        "swd_69248",
        "swd_69347",
        "swd_69389",
        "swd_69406",
        "swd_69410",
        "swd_69441",
        "swd_69498",
        "swd_69544",
        "swd_69547",
        "swd_69550",
        "swd_69730",
        "swd_69740",
        "swd_69770",
        "swd_69815",
        "swd_69832",
        "swd_69917",
        
        "swd_70118",
        "swd_70168",
        "swd_70260",
        "swd_70262",
        "swd_70299",
        "swd_70378",
        "swd_70401",
        "swd_70403",
        "swd_70451",
        "swd_70471",
        "swd_70557",
        "swd_70558",
        "swd_70560",
        "swd_70569",
        "swd_70570",
        "swd_70574",
        "swd_70577",
        "swd_70586",
        "swd_70599",
        "swd_70613",
        "swd_70614",
        "swd_70676",
        "swd_70701",
        "swd_70704",
        "swd_70709",
        "swd_70710",
        "swd_70737",
        "swd_70738",
        "swd_70739",
        "swd_70740",
        "swd_70744",
        "swd_70750",
        "swd_70753",
        "swd_70761",
        "swd_70769",
        "swd_70774",
        "swd_70779",
        "swd_70792",
        "swd_70830",
        "swd_70845",
        "swd_70850",
        "swd_70864",
        "swd_70877",
        "swd_70887",
        "swd_70894",
        "swd_70900",
        "swd_70912",
        "swd_70915",
        "swd_70916",
        "swd_70917",
        "swd_70932",
        "swd_70936",
        "swd_70937",
        "swd_70939",
        "swd_70941",
        "swd_70944",
        "swd_70948",
        "swd_70959",
        "swd_70963",
        "swd_70966",
        "swd_70967",
        "swd_70971",
        "swd_70974",
        "swd_70976",
        "swd_70980",
        "swd_70986",
        "swd_70987",
        "swd_70992",
        "swd_70994",
        "swd_70995",
        "swd_71006",
        "swd_71011",
        "swd_71019",
        "swd_71025",
        "swd_71029",
        "swd_71031",
        "swd_71032",
        "swd_71034",
        "swd_71041",
        "swd_71042",
        "swd_71046",
        "swd_71048",
        "swd_71051",
        "swd_71052",
        "swd_71053",
        "swd_71055",
        "swd_71056",
        "swd_71059",
        "swd_71063",
        "swd_71071",
        "swd_71078",
        "swd_71090",
        "swd_71092",
        "swd_71096",
        "swd_71103",
        "swd_71106",
        "swd_71115",
        "swd_71116",
        "swd_71117",
        "swd_71119",
        "swd_71121",
        "swd_71122",
        "swd_71125",
        "swd_71128",
        "swd_71142",
        "swd_71143",
        "swd_71144",
        "swd_71149",
        "swd_71150",
        "swd_71152",
        "swd_71162",
        "swd_71165",
        "swd_71168",
        "swd_71173",
        "swd_71175",
        "swd_71178",
        "swd_71201",
        "swd_71203",
        "swd_71204",
        "swd_71205",
        "swd_71261",
        "swd_71269",
        "swd_71275",
        "swd_71277",
        "swd_71279",
        "swd_71282",
        "swd_71285",
        "swd_71286",
        "swd_71307",
        "swd_71319",
        "swd_71585",
        "swd_71596",
        "swd_71606",
        "swd_71608",
        "swd_71625",
        "swd_71631",
        "swd_71636",
        "swd_71683",
        "swd_71687",
        "swd_71751",
        "swd_71754",
        "swd_71756",
        "swd_71929",
        "swd_72027",
        "swd_72030",
        "swd_72031",
        "swd_72053",
        "swd_72054",
        "swd_72057",
        "swd_72059",
        "swd_72061",
        "swd_72062",
        "swd_72063",
        "swd_72069",
        "swd_72073",
        "swd_72074",
        "swd_72078",
        "swd_72084",
        "swd_72093",
        "swd_72108",
        "swd_72117",
        "swd_72188",
        "swd_72189",
        "swd_72190",
        "swd_72207",
        "swd_72285",
        "swd_72328",
        "swd_72335",
        "swd_72340",
        "swd_72344",
        "swd_72350",
        "swd_72351",
        "swd_72357",
        "swd_72360",
        "swd_72361",
        "swd_72362",
        "swd_72370",
        "swd_72373",
        "swd_72390",
        "swd_72396",
        "swd_72405",
        "swd_72406",
        "swd_72412",
        "swd_72419",
        "swd_72425",
        "swd_72428",
        "swd_72429",
        "swd_72433",
        "swd_72436",
        "swd_72440",
        "swd_72446",
        "swd_72449",
        "swd_72454",
        "swd_72455",
        "swd_72459",
        "swd_72461",
        "swd_72494",
        "swd_72495",
        "swd_72610",
        "swd_72611",
        "swd_72620",
        "swd_72621",
        "swd_72622",
        "swd_72646",
        "swd_72647",
        "swd_72656",
        "swd_72663",
        "swd_72668",
        "swd_72669",
        "swd_72673",
        "swd_72680",
        "swd_72684",
        "swd_72686",
        "swd_72688",
        "swd_72689",
        "swd_72693",
        "swd_72699",
        "swd_72701",
        "swd_72704",
        "swd_72709",
        "swd_72720",
        "swd_72723",
        "swd_72724",
        "swd_72733",
        "swd_72735",
        "swd_72738",
        "swd_72740",
        "swd_72742",
        "swd_72745",
        "swd_72747",
        "swd_72748",
        "swd_72752",
        "swd_72762",
        "swd_72818",
        "swd_72819",
        "swd_72826",
        "swd_72838",
        "swd_72850",
        "swd_72852",
        "swd_72857",
        "swd_72876",
        "swd_72893",
        "swd_72895",
        "swd_72897",

# ... sar ---------------------------------------------------------------------
        "sar_166",
        "sar_444",
        "sar_460",
        "sar_464",
        "sar_477",
        
        "sar_3319",
        "sar_3976",
        "sar_3978",
        "sar_3981",
        
        "sar_4873",
        
        "sar_5250",
        "sar_5482",
        "sar_5483",
        "sar_5487",
        "sar_5500",
        "sar_5507",
        "sar_5508",
        "sar_5510",
        "sar_5511",
        "sar_5513",
        "sar_5514",
        "sar_5516",
        "sar_5517",
        "sar_5521",
        "sar_5523",
        "sar_5527",
        "sar_5528",
        "sar_5529",
        "sar_5530",
        "sar_5531",
        "sar_5539",
        "sar_5541",
        "sar_5542",
        #"sar_5543",
        "sar_5545",
        "sar_5546",
        "sar_5555",
        "sar_5822",
        
        "sar_6536",
        "sar_6547",
        
        "sar_7129",
        "sar_7140",
        "sar_7285",
        "sar_7288",
        "sar_7289",
        "sar_7307",
        "sar_7318",
        "sar_7724",
        
        "sar_8077",
        "sar_8079",
        "sar_8081",
        "sar_8083",
        "sar_8086",
        "sar_8087",
        "sar_8089",
        "sar_8090",
        "sar_8095",
        "sar_8098",
        "sar_8099",
        "sar_8100",
        "sar_8102",
        "sar_8104",
        "sar_8108",
        "sar_8112",
        "sar_8113",
        "sar_8115",
        "sar_8118",
        "sar_8120",
        "sar_8122",
        "sar_8124",
        "sar_8131",
        "sar_8132",
        "sar_8134",
        "sar_8135",
        "sar_8136",
        "sar_8137",
        "sar_8140",
        "sar_8141",
        "sar_8142",
        "sar_8147",
        "sar_8148",
        "sar_8149",
        "sar_8150",
        "sar_8152",
        "sar_8154",
        "sar_8156",
        "sar_8159",
        "sar_8160",
        "sar_8165",
        "sar_8167",
        "sar_8170",
        "sar_8171",
        "sar_8248"
))



#  SPLIT LINES  -----------------------------------------------------------
print("### --- SPLIT SEGMENTS --- ###")
options(warn = -1)
data = st_as_sf(dt_rivers)
data = split_lines(data  = data, 
                   split = "vdn_7815", 
                   by    = "sar_8075")
# save original rlp10112 to split vdn_2668 with it
splitter = filter(data, ecoserv_id == "rlp_10112")
data = split_lines(data  = data, 
                   split = "rlp_10112", 
                   by    = "vdn_2668")
data = split_lines_external(data = data,
                            split = "vdn_2668",
                            by = splitter)


options(warn = 1)
setDT(data)
dt_rivers = data

# modify added lines  -----------------------------------------------------
dt_rivers[ecoserv_id == "add_72929", FROM := "P103941"]
dt_rivers[ecoserv_id == "split_1", FROM := "P127267"]
dt_rivers[ecoserv_id == "split_1", TO := "P29431"]
dt_rivers[ecoserv_id == "split_2", FROM := "P19952"]
dt_rivers[ecoserv_id == "split_2", TO := "P2582"]
dt_rivers[ecoserv_id == "rlp_10112", FROM := "P2580"]
dt_rivers[ecoserv_id == "rlp_10112", TO := "P19952"]
dt_rivers[ecoserv_id == "vdn_7815", FROM := "P29518"]
dt_rivers[ecoserv_id == "vdn_7815", TO := "P127267"]
dt_rivers[ecoserv_id == "vdn_2668", FROM := "P19951"]
dt_rivers[ecoserv_id == "vdn_2668", TO := "P19952"]

## ---- add rows ----- ## 
new_number_rlp <- dt_rivers[str_detect(string=ecoserv_id,pattern="rlp"), max(ecoserv_number)]
dt_new_row <- data.table(ecoserv_id = paste0("rlp_", new_number_rlp + 1:2), 
                         FROM       = c("P42","P52"   ),
                         TO         = c("P44", "P2183"), 
                         score_up   = 1,
                         score_down = 1,
                         ecoserv_number = new_number_rlp + 1:2, 
                         geom       = c(dt_rivers[TO == "P42", geom], 
                                        dt_rivers[TO == "P52", geom])
)

dt_rivers <- rbindlist(list(dt_rivers, dt_new_row), 
                       use.names = TRUE)

## second delete step 
dt_rivers = dt_rivers[!ecoserv_id %in% c(
        "rlp_10404",
        "rlp_16932",
        "rlp_16934",
        "rlp_16961",
        "rlp_16962",
        # temp 
        "rlp_10138"
)
]

rm(dt_new_row, new_number_rlp, reverse);gc()




# SAVE  -----------------------------------------------------------
if (OPTIONS$save) {
        saveRDS(object=dt_rivers,
                file=file.path(DIR$da, "fixed_rivers.RDS"))
}
if (OPTIONS$remove) {
        rm(list=ls());gc()
}
# WORKSHOP ---------------------------------------------------------------
if (OPTIONS$workshop){
        
        tmap_mode("view")
        st_rivers = st_as_sf(dt_rivers)
        st_sites%<>%st_transform(crs=st_crs(st_rivers))
        
        # ## -- what? -- ## 
        sub=tb_edge_id_sub$ecoserv_id
        
        sub=c("PW_01", "PW02")
        
        rivers_sub=dt_rivers[ecoserv_id %in% sub]
        
        rivers_sub %>%
                mutate(fromto=paste(FROM, TO)) %>%
                st_as_sf() %>%
                tm_shape() + tm_lines(col="ecoserv_id", lwd=3) +
                tm_text(text="fromto")
        # 
        # ## -- bbox around sites -- ## 
        # ## -- + plot            -- ## 
        st_sites_sub = filter(st_sites, site %in% c("PW_01", "PW_02"))
        
        st_sites_sub %>%
                st_bbox() * c(.99,.99,1.11,1.11) ->
                cut_bbox
        # 
        st_cropped=sf::st_crop(
                x=st_rivers,
                y=cut_bbox
        )
        # 
        tm_shape(st_cropped) + tm_lines() +
                tm_shape(st_sites_sub) + tm_dots(col = "red")
        # 
        # ## ------------------- ## 
        # ## -- add lines eda -- ## 
        
        # extract lines between which a link is missing 
        st_rivers_sub =
                st_rivers %>%
                filter(ecoserv_id %in% c("rlp_188", "rlp_9913"))
        
        st_data = st_rivers
        st_data_sub = st_data %>%
                filter(ecoserv_id %in% c(from_line, to_line))
        
        st_data_sub_points =
                st_data_sub %>%
                st_cast(to = "POINT") %>%
                mutate(p_id = row_number())
        
        st_line_to = st_data_sub %>%
                filter(ecoserv_id == to_line)
        st_line_from = st_data_sub %>%
                filter(ecoserv_id == from_line)
        st_points_from = st_data_sub_points %>%
                filter(ecoserv_id == from_line)
        st_points_to = st_data_sub_points %>%
                filter(ecoserv_id == to_line)
        st_points_from = st_points_from[c(1, nrow(st_points_from)), ]
        st_points_to = st_points_to[c(1, nrow(st_points_to)), ]
        
        tm_shape(st_points_from) + tm_dots() +
                tm_shape(st_points_to) + tm_dots()
        
        p_id1 = st_nearest_feature(x = st_line_to,
                                   y = st_points_from)
        p_id2 = st_nearest_feature(x = st_line_from,
                                   y = st_points_to)
        p_id1 = st_points_from$p_id[p_id1]
        p_id2 = st_points_to$p_id[p_id2]
        
        # cast to points
        options(warn=-1);
        st_rivers_sub_point =
                st_rivers_sub %>%
                st_cast(to = "POINT");
        options(warn=1)
        
        st_rivers_sub_point %<>%
                mutate(p_id = 1:nrow(st_rivers_sub_point))
        point_id_2 <-
                unique(st_nearest_feature(
                        filter(st_rivers_sub_point, ecoserv_id == "rlp_188"),
                        filter(st_rivers_sub_point, ecoserv_id == "rlp_9913")
                ))
        point_id_1 <-
                unique(st_nearest_feature(
                        filter(st_rivers_sub_point, ecoserv_id == "rlp_9913"),
                        filter(st_rivers_sub_point, ecoserv_id == "rlp_188")
                ))
        point_id_1 = filter(st_rivers_sub_point, ecoserv_id == "rlp_9913") %>%
                st_drop_geometry() %>%
                select(p_id) %>%
                .[point_id_1,]
        
        point_id_2 = filter(st_rivers_sub_point, ecoserv_id == "rlp_188") %>%
                st_drop_geometry() %>%
                select(p_id) %>%
                .[point_id_2,]
        
        tm_shape(st_rivers_sub_point) +
                tm_dots()
        
        st_rivers_sub_point %>%
                filter(p_id %in% c(442, 446)) %>%
                st_coordinates(st_new_line) %>%
                .[, -3] %>%
                list() %>%
                st_multilinestring() %>%
                st_sfc() ->
                st_new_line
        
        st_crs(st_new_line)=st_crs(st_rivers_sub)
        
        tm_shape(st_rivers_sub) + tm_lines() +
                tm_shape(st_new_line) + tm_lines()
        
        df_attributes = data.frame(
                "ecoserv_id" = paste0("add_",
                                      max(dt_rivers$ecoserv_number) + 1),
                "FROM" = "P40",
                "TO"   = "P44",
                score_up=1,
                score_down=1,
                ecoserv_number=max(dt_rivers$ecoserv_number) + 1
        )
        
        st_new_line = st_sf(df_attributes, geom = st_new_line)
        st_crs(st_new_line)=st_crs(st_rivers)
        st_rivers %<>% rbind(st_new_line)
        
        
        #-----------------------#
        
        
        
        ## ------------------- ##
        
        
        to_split <- rivers[ecoserv_id == "vdn_7861"]
        to_split %<>% st_as_sf()
        to_split %<>% st_cast("POINT")
        
        st_nearest_feature(y = to_split,
                           x = st_as_sf(rivers[ecoserv_id == "rlp_16958"]))
}
